+++ src/keyed/vec.rs
@@ -175,6 +175,17 @@ impl<K: Key, V, A: Alloc> core::ops::DerefMut for KVec<K, V, A> {
     }
 }
 
+
+impl<K: Key, V: PartialEq, A: Alloc> PartialEq for KVec<K, V, A> {
+    #[inline(always)]
+    fn eq(&self, other: &Self) -> bool {
+        self.as_slice() == other.as_slice()
+    }
+}
+
+impl<K: Key, V: Eq, A: Alloc> Eq for KVec<K, V, A> {}
+
+
 impl<K: Key, V, A: Alloc> core::fmt::Debug for KVec<K, V, A>
 where K: core::fmt::Debug, V: core::fmt::Debug {
     fn fmt(&self, f: &mut core::fmt::Formatter) -> core::fmt::Result {